workflow: update migration script
